enh: add gpu utilization info
